Miscellaneous refactoring.

+++ BigDecimalTests/TestToString.cs
@@ -16,19 +16,19 @@ public class TestToString
         str = bd.ToString("D");
         Assert.AreEqual("12345", str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("D");
         Assert.AreEqual("12345E+67", str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("D");
         Assert.AreEqual("12345E-67", str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("D");
         Assert.AreEqual("-12345E+67", str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("D");
         Assert.AreEqual("-12345E-67", str);
 
@@ -49,19 +49,19 @@ public class TestToString
         str = bd.ToString("DU");
         Assert.AreEqual("12345", str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("DU");
         Assert.AreEqual("12345×10⁶⁷", str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("DU");
         Assert.AreEqual("12345×10⁻⁶⁷", str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("DU");
         Assert.AreEqual("-12345×10⁶⁷", str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("DU");
         Assert.AreEqual("-12345×10⁻⁶⁷", str);
 
@@ -81,19 +81,19 @@ public class TestToString
         str = bd.ToString("E");
         Assert.AreEqual("1.2345E+004", str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("E");
         Assert.AreEqual("1.2345E+071", str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("E");
         Assert.AreEqual("1.2345E-063", str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("E");
         Assert.AreEqual("-1.2345E+071", str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("E");
         Assert.AreEqual("-1.2345E-063", str);
 
@@ -109,7 +109,7 @@ public class TestToString
         BigDecimal bd;
         string str;
 
-        bd = 1.2345;
+        bd = (BigDecimal)1.2345;
         str = bd.ToString("E2");
         Assert.AreEqual("1.23E+000", str);
 
@@ -117,19 +117,19 @@ public class TestToString
         str = bd.ToString("E2");
         Assert.AreEqual("1.23E+004", str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)(12345e67);
         str = bd.ToString("E2");
         Assert.AreEqual("1.23E+071", str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)(12345e-67);
         str = bd.ToString("E2");
         Assert.AreEqual("1.23E-063", str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("E2");
         Assert.AreEqual("-1.23E+071", str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("E2");
         Assert.AreEqual("-1.23E-063", str);
 
@@ -148,19 +148,19 @@ public class TestToString
         str = bd.ToString("EU");
         Assert.AreEqual("1.2345×10" + "4".ToSuperscript(), str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("EU");
         Assert.AreEqual("1.2345×10" + "71".ToSuperscript(), str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("EU");
         Assert.AreEqual("1.2345×10" + "-63".ToSuperscript(), str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("EU");
         Assert.AreEqual("-1.2345×10" + "71".ToSuperscript(), str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("EU");
         Assert.AreEqual("-1.2345×10" + "-63".ToSuperscript(), str);
 
@@ -176,7 +176,7 @@ public class TestToString
         BigDecimal bd;
         string str;
 
-        bd = 1.2345;
+        bd = (BigDecimal)1.2345;
         str = bd.ToString("E2U");
         Assert.AreEqual("1.23×10" + "0".ToSuperscript(), str);
 
@@ -184,19 +184,19 @@ public class TestToString
         str = bd.ToString("E2U");
         Assert.AreEqual("1.23×10" + "4".ToSuperscript(), str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("E2U");
         Assert.AreEqual("1.23×10" + "71".ToSuperscript(), str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("E2U");
         Assert.AreEqual("1.23×10" + "-63".ToSuperscript(), str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("E2U");
         Assert.AreEqual("-1.23×10" + "71".ToSuperscript(), str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("E2U");
         Assert.AreEqual("-1.23×10" + "-63".ToSuperscript(), str);
 
@@ -256,11 +256,11 @@ public class TestToString
         BigDecimal bd;
         string str;
 
-        bd = 1.2345;
+        bd = (BigDecimal)1.2345;
         str = bd.ToString("F3");
         Assert.AreEqual("1.235", str);
 
-        bd = 123.45;
+        bd = (BigDecimal)123.45;
         str = bd.ToString("F3");
         Assert.AreEqual("123.450", str);
 
@@ -268,27 +268,27 @@ public class TestToString
         str = bd.ToString("F3");
         Assert.AreEqual("12345.000", str);
 
-        bd = 12345e2;
+        bd = (BigDecimal)12345e2;
         str = bd.ToString("F3");
         Assert.AreEqual("1234500.000", str);
 
-        bd = 1.2345e7;
+        bd = (BigDecimal)1.2345e7;
         str = bd.ToString("F3");
         Assert.AreEqual("12345000.000", str);
 
-        bd = 1.2345e-2;
+        bd = (BigDecimal)1.2345e-2;
         str = bd.ToString("F3");
         Assert.AreEqual("0.012", str);
 
-        bd = 12345e-6;
+        bd = (BigDecimal)12345e-6;
         str = bd.ToString("F3");
         Assert.AreEqual("0.012", str);
 
-        bd = -12345e2;
+        bd = (BigDecimal)(-12345e2);
         str = bd.ToString("F3");
         Assert.AreEqual("-1234500.000", str);
 
-        bd = -12345e-6;
+        bd = (BigDecimal)(-12345e-6);
         str = bd.ToString("F3");
         Assert.AreEqual("-0.012", str);
 
@@ -303,15 +303,15 @@ public class TestToString
         BigDecimal bd;
         string str;
 
-        bd = 1.2345;
+        bd = (BigDecimal)1.2345;
         str = bd.ToString("F0");
         Assert.AreEqual("1", str);
 
-        bd = 12345.5;
+        bd = (BigDecimal)12345.5;
         str = bd.ToString("F0");
         Assert.AreEqual("12346", str);
 
-        bd = 123.45;
+        bd = (BigDecimal)123.45;
         str = bd.ToString("F0");
         Assert.AreEqual("123", str);
 
@@ -319,27 +319,27 @@ public class TestToString
         str = bd.ToString("F0");
         Assert.AreEqual("12345", str);
 
-        bd = 12345e2;
+        bd = (BigDecimal)12345e2;
         str = bd.ToString("F0");
         Assert.AreEqual("1234500", str);
 
-        bd = 1.2345e7;
+        bd = (BigDecimal)1.2345e7;
         str = bd.ToString("F0");
         Assert.AreEqual("12345000", str);
 
-        bd = 1.2345e-2;
+        bd = (BigDecimal)1.2345e-2;
         str = bd.ToString("F0");
         Assert.AreEqual("0", str);
 
-        bd = 12345e-6;
+        bd = (BigDecimal)12345e-6;
         str = bd.ToString("F0");
         Assert.AreEqual("0", str);
 
-        bd = -12345e2;
+        bd = (BigDecimal)(-12345e2);
         str = bd.ToString("F0");
         Assert.AreEqual("-1234500", str);
 
-        bd = -12345e-6;
+        bd = (BigDecimal)(-12345e-6);
         str = bd.ToString("F0");
         Assert.AreEqual("0", str);
 
@@ -358,7 +358,7 @@ public class TestToString
         BigDecimal bd;
         string str;
 
-        bd = 1.2345;
+        bd = (BigDecimal)1.2345;
         str = bd.ToString("G");
         Assert.AreEqual("1.2345", str);
 
@@ -366,43 +366,43 @@ public class TestToString
         str = bd.ToString("G");
         Assert.AreEqual("12345", str);
 
-        bd = 12345e2;
+        bd = (BigDecimal)12345e2;
         str = bd.ToString("G");
         Assert.AreEqual("1234500", str);
 
-        bd = 1.2345e7;
+        bd = (BigDecimal)1.2345e7;
         str = bd.ToString("G");
         Assert.AreEqual("12345000", str);
 
-        bd = 1.2345e-2;
+        bd = (BigDecimal)1.2345e-2;
         str = bd.ToString("G");
         Assert.AreEqual("0.012345", str);
 
-        bd = 12345e-6;
+        bd = (BigDecimal)12345e-6;
         str = bd.ToString("G");
         Assert.AreEqual("0.012345", str);
 
-        bd = -12345e2;
+        bd = (BigDecimal)(-12345e2);
         str = bd.ToString("G");
         Assert.AreEqual("-1234500", str);
 
-        bd = -12345e-6;
+        bd = (BigDecimal)(-12345e-6);
         str = bd.ToString("G");
         Assert.AreEqual("-0.012345", str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("G");
         Assert.AreEqual("1.2345E+71", str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("G");
         Assert.AreEqual("1.2345E-63", str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("G");
         Assert.AreEqual("-1.2345E+71", str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("G");
         Assert.AreEqual("-1.2345E-63", str);
 
@@ -419,7 +419,7 @@ public class TestToString
         BigDecimal bd;
         string str;
 
-        bd = 1.2345;
+        bd = (BigDecimal)1.2345;
         str = bd.ToString("G3");
         Assert.AreEqual("1.23", str);
 
@@ -427,43 +427,43 @@ public class TestToString
         str = bd.ToString("G3");
         Assert.AreEqual("12300", str);
 
-        bd = 12345e2;
+        bd = (BigDecimal)12345e2;
         str = bd.ToString("G3");
         Assert.AreEqual("1230000", str);
 
-        bd = 1.2345e7;
+        bd = (BigDecimal)1.2345e7;
         str = bd.ToString("G3");
         Assert.AreEqual("12300000", str);
 
-        bd = 1.2345e-2;
+        bd = (BigDecimal)1.2345e-2;
         str = bd.ToString("G3");
         Assert.AreEqual("0.0123", str);
 
-        bd = 12345e-6;
+        bd = (BigDecimal)12345e-6;
         str = bd.ToString("G3");
         Assert.AreEqual("0.0123", str);
 
-        bd = -12345e2;
+        bd = (BigDecimal)(-12345e2);
         str = bd.ToString("G3");
         Assert.AreEqual("-1230000", str);
 
-        bd = -12345e-6;
+        bd = (BigDecimal)(-12345e-6);
         str = bd.ToString("G3");
         Assert.AreEqual("-0.0123", str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("G3");
         Assert.AreEqual("1.23E+71", str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("G3");
         Assert.AreEqual("1.23E-63", str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("G3");
         Assert.AreEqual("-1.23E+71", str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("G3");
         Assert.AreEqual("-1.23E-63", str);
 
@@ -478,7 +478,7 @@ public class TestToString
         BigDecimal bd;
         string str;
 
-        bd = 1.2345;
+        bd = (BigDecimal)1.2345;
         str = bd.ToString("GU");
         Assert.AreEqual("1.2345", str);
 
@@ -486,43 +486,43 @@ public class TestToString
         str = bd.ToString("GU");
         Assert.AreEqual("12345", str);
 
-        bd = 12345e2;
+        bd = (BigDecimal)12345e2;
         str = bd.ToString("GU");
         Assert.AreEqual("1234500", str);
 
-        bd = 1.2345e7;
+        bd = (BigDecimal)1.2345e7;
         str = bd.ToString("GU");
         Assert.AreEqual("12345000", str);
 
-        bd = 1.2345e-2;
+        bd = (BigDecimal)1.2345e-2;
         str = bd.ToString("GU");
         Assert.AreEqual("0.012345", str);
 
-        bd = 12345e-6;
+        bd = (BigDecimal)12345e-6;
         str = bd.ToString("GU");
         Assert.AreEqual("0.012345", str);
 
-        bd = -12345e2;
+        bd = (BigDecimal)(-12345e2);
         str = bd.ToString("GU");
         Assert.AreEqual("-1234500", str);
 
-        bd = -12345e-6;
+        bd = (BigDecimal)(-12345e-6);
         str = bd.ToString("GU");
         Assert.AreEqual("-0.012345", str);
 
-        bd = 12345e67;
+        bd = (BigDecimal)12345e67;
         str = bd.ToString("GU");
         Assert.AreEqual("1.2345×10⁷¹", str);
 
-        bd = 12345e-67;
+        bd = (BigDecimal)12345e-67;
         str = bd.ToString("GU");
         Assert.AreEqual("1.2345×10⁻⁶³", str);
 
-        bd = -12345e67;
+        bd = (BigDecimal)(-12345e67);
         str = bd.ToString("GU");
         Assert.AreEqual("-1.2345×10⁷¹", str);
 
-        bd = -12345e-67;
+        bd = (BigDecimal)(-12345e-67);
         str = bd.ToString("GU");
         Assert.AreEqual("-1.2345×10⁻⁶³", str);
