+++ BigDecimalTests/TestTrig.cs
@@ -91,7 +91,7 @@ public class TestTrig
                 strAsinD = "0.00000000000000";
             }
 
-            BigDecimal bd = d;
+            BigDecimal bd = (BigDecimal)d;
             BigDecimal asinBD = BigDecimal.Asin(bd);
             string strAsinBD = asinBD.ToString("F14");
 
@@ -156,7 +156,7 @@ public class TestTrig
                 strAcosD = "0.00000000000000";
             }
 
-            BigDecimal bd = d;
+            BigDecimal bd = (BigDecimal)d;
             BigDecimal acosBD = BigDecimal.Acos(bd);
             string strAcosBD = acosBD.ToString("F14");
 
@@ -178,7 +178,7 @@ public class TestTrig
                 strAtanD = "0.00000000000000";
             }
 
-            BigDecimal bd = d;
+            BigDecimal bd = (BigDecimal)d;
             BigDecimal atanBD = BigDecimal.Atan(bd);
             string strAtanBD = atanBD.ToString("F14");
